Squashed commit of the following:
commit 16baa544b7e8af7443ca1f2ea01c7c5d2cea3d29
Author: Ariansdf <38947447+Ariansdf@users.noreply.github.com>
Date:   Thu Sep 11 13:48:43 2025 +0200

    FineTune

commit e1fa690bfd009ed4ae8424b59f7ec7309fd4a2e8
Author: Ariansdf <38947447+Ariansdf@users.noreply.github.com>
Date:   Wed Sep 10 19:34:22 2025 +0200

    avatar Migration updated

commit 7f1d1a1b201dad9e8789f0ab174096a230391b70
Author: Ariansdf <38947447+Ariansdf@users.noreply.github.com>
Date:   Wed Sep 10 19:18:18 2025 +0200

    backup: download feature, bugfixes

commit e661f6ec8eaa63d85a049a749e3117f7dfcd7de1
Author: Ariansdf <38947447+Ariansdf@users.noreply.github.com>
Date:   Wed Sep 10 11:43:49 2025 +0200

    Finetune Styling

commit d01bbc7edf98cb913809b577cc472797900fd2a1
Author: Ariansdf <38947447+Ariansdf@users.noreply.github.com>
Date:   Tue Sep 9 18:29:27 2025 +0200

    Typo Correction public_key

commit fb08b50b949408680bca5389e13f9de6c6811a45
Author: Ariansdf <38947447+Ariansdf@users.noreply.github.com>
Date:   Tue Sep 9 15:59:49 2025 +0200

    Auto Basic Announcement Migration

commit e608d810953c4f64b8f265cade95594379010300
Author: Ariansdf <38947447+Ariansdf@users.noreply.github.com>
Date:   Tue Sep 9 14:14:36 2025 +0200

    Garbage Collection added to FileStorage

commit b8f7bc11c00a485281cbaf9b2969b7c1a79f6783
Author: Ariansdf <38947447+Ariansdf@users.noreply.github.com>
Date:   Tue Sep 9 13:12:31 2025 +0200

    Correction

commit 43ac059748e944ee8a579340af329230d750f418
Author: Ariansdf <38947447+Ariansdf@users.noreply.github.com>
Date:   Tue Sep 9 13:12:21 2025 +0200

    HAWKI User Adjustment

    Seeder converted to migration
    Avatar added to public folder and automated into storage
    Avatar Update Command added for quality of life

commit a6cd3411a10f35d8630ea2bf14c43ebde81dade9
Author: Ariansdf <38947447+Ariansdf@users.noreply.github.com>
Date:   Tue Sep 9 11:00:36 2025 +0200

    Bugfix: 419 After Auto Logout

commit bbf36edc292b09c68f109812bdc8b4c34e716d88
Author: Ariansdf <38947447+Ariansdf@users.noreply.github.com>
Date:   Tue Sep 9 10:23:07 2025 +0200

    add accept json to fetch requests

commit 2a9ae6618b64c8a3002037c71017e0451f6be452
Author: Ariansdf <38947447+Ariansdf@users.noreply.github.com>
Date:   Mon Sep 8 18:15:14 2025 +0200

    Gwdg Docling added

commit 8fc3e61afb557d8d490f6c075d4b20e3f22c31fa
Author: Ariansdf <38947447+Ariansdf@users.noreply.github.com>
Date:   Mon Sep 8 15:16:18 2025 +0200

    file storage structure update

commit d4504cb32b621a5578d2673c27bd28a5ba2bca0b
Author: Ariansdf <38947447+Ariansdf@users.noreply.github.com>
Date:   Fri Sep 5 16:02:58 2025 +0200

    cleanup

commit 2c8f03ff7b7e121c689ffd2faf84329cef7ac20f
Merge: 3732cb0 413e113
Author: Ariansdf <38947447+Ariansdf@users.noreply.github.com>
Date:   Fri Sep 5 14:40:09 2025 +0200

    Merge branch 'feature/2.0.1_f2' into feature/DEV-Unfinished

    # Conflicts:
    #	app/Http/Controllers/ProfileController.php
    #	public/css_v2.0.1_f1/chat_modules.css
    #	resources/language/de_DE.json
    #	resources/views/layouts/home.blade.php
    #	resources/views/partials/gateway/register.blade.php

commit 3732cb0e56bca22af7508336a2ab16f6b498a478
Author: Ariansdf <38947447+Ariansdf@users.noreply.github.com>
Date:   Fri Sep 5 14:34:10 2025 +0200

    fix problem in groupchat regeneration

commit 79540580c2b9639f884251307c1556604121217e
Merge: d0d6721 123f133
Author: Ariansdf <38947447+Ariansdf@users.noreply.github.com>
Date:   Fri Sep 5 11:58:28 2025 +0200

    Merge branch 'feature/maybreakchanges' into feature/DEV-Unfinished

    # Conflicts:
    #	app/Http/Controllers/StreamController.php
    #	app/Services/Chat/Room/Traits/RoomMessages.php
    #	public/js_v2.0.1_f1/groupchat_functions.js

commit d0d6721c0091015a9499524d0c20986b52461d1a
Author: Ariansdf <38947447+Ariansdf@users.noreply.github.com>
Date:   Fri Sep 5 11:51:36 2025 +0200

    BACKUP:: Large Commit, Refactor AI Conn

commit 123f13355a8af369e929bba1ecc260d50abc056b
Author: Ariansdf <38947447+Ariansdf@users.noreply.github.com>
Date:   Tue Sep 2 10:37:08 2025 +0200

    profileService traits to classes

commit eb638f0280939d700d3935f0572c28994cda8592
Author: Ariansdf <38947447+Ariansdf@users.noreply.github.com>
Date:   Tue Sep 2 10:19:52 2025 +0200

    Solved Error for large boardcast packets.

commit 413e113c532f515d8d4b2fca3debba954a9d50d2
Author: Ariansdf <38947447+Ariansdf@users.noreply.github.com>
Date:   Tue Apr 29 15:17:21 2025 +0200

    Passkey update

+++ app/Services/Chat/Message/Handlers/PrivateMessageHandler.php
@@ -6,33 +6,24 @@ namespace App\Services\Chat\Message\Handlers;
 use App\Models\User;
 use App\Models\AiConv;
 use App\Models\AiConvMsg;
-
-
+use App\Models\Room;
+use App\Services\Chat\Attachment\AttachmentService;
+use Illuminate\Auth\Access\AuthorizationException;
 use Illuminate\Support\Facades\Auth;
-use Illuminate\Support\Facades\Log;
-use Illuminate\Support\Facades\Storage;
-
-use Illuminate\Http\Request;
-
 
 class PrivateMessageHandler extends BaseMessageHandler{
 
 
 
-    public function create(array $data, string $slug): array {
-
-        $conv = AiConv::where('slug', $slug)->firstOrFail();
+    public function create(AiConv|Room $conv, array $data): AiConvMsg {
         if ($conv->user_id !== Auth::id()) {
-            return [
-                        'success'=> false,
-                        'error' => 'Access denied'
-            ];
+            throw new AuthorizationException();
         }
 
         $user = $data['isAi'] ? User::find(1) : Auth::user();
         $messageRole = $data['isAi'] ? 'assistant' : 'user';
 
-        $nextMessageId = $this->assignID($conv, $data['threadID']);
+        $nextMessageId = $this->assignID($conv, $data['threadId']);
         $message = AiConvMsg::create([
             'conv_id' => $conv->id,
             'user_id' => $user->id,
@@ -52,24 +43,18 @@ class PrivateMessageHandler extends BaseMessageHandler{
             if($attachments){
                 foreach($attachments as $attach){
                     $this->attachmentService->assignToMessage($message, $attach);
+
                 }
             }
         }
-
-        $messageData = $message->createMessageObject();
-        return [
-            'success'=> true,
-            'messageData'=> $messageData
-        ];
+        return $message;
     }
 
 
 
-    public function update(array $data, string $slug){
-
-        $conv = AiConv::where('slug', $slug)->firstOrFail();
+    public function update(AiConv|Room $conv, array $data): AiConvMsg{
         if ($conv->user_id !== Auth::id()) {
-            return response()->json(['error' => 'Access denied'], 403);
+            throw new AuthorizationException();
         }
 
         //find the target message
@@ -83,27 +68,27 @@ class PrivateMessageHandler extends BaseMessageHandler{
             'completion' => $data['completion']
         ]);
 
-        $messageData = $message->toArray();
-        $messageData['created_at'] = $message->created_at->format('Y-m-d+H:i');
-        $messageData['updated_at'] = $message->updated_at->format('Y-m-d+H:i');
-
-        return $messageData;
-
+        return $message;
     }
 
 
-    public function delete(array $data, string $slug){
-
-        $conv = AiConv::where('slug', $slug)->firstOrFail();
+    public function delete(AiConv|Room $conv, array $data): bool{
         if ($conv->user_id !== Auth::id()) {
-            return response()->json([
-                'success'=> false,
-                'error' => 'Access denied'
-            ], 403);
+            throw new AuthorizationException();
         }
 
         $message = $conv->messages->where('message_id', $data['message_id'])->first();
-        $message->delete();
+        if (!$message->user->is(Auth::user())) {
+            throw new AuthorizationException();
+        }
 
+        $attachmentService = app(AttachmentService::class);
+        $attachments = $message->attachments;
+        foreach ($attachments as $attachment) {
+            $attachmentService->delete($attachment);
+        }
+
+        $message->delete();
+        return true;
     }
 }
